Walking Audio, Invisible Walls, Dialogue Movement, Dialogue Fix
A lot of things happened in this new update. The walking audio is now fixed with each individual footstep instead of a looping audio clip. I've also added invisible walls around the level so that people can't just walk off the map. I've fixed the error with the Grandma quest. Movement is also restricted when in dialogue with NPCs.